Seni made function in PassengerDBUtil to return arraylist of passenger info
so I have to next make use of this function to display customers info. when the customer login or register s/he is directed to homepage from that point in that homepage.jsp if I press account button i must use session to get pid of that passenger and display the account info. not sure which servlet to use. have to ask sai to use tht if function so i can use same servlet instead of making another one